feat: hide `suggestions shelf`
